requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
black==25.1.0
click==8.1.8
dj-database-url==2.3.0
Django==4.2.18
django-constance==4.3.1
django-cors-headers==3.13.0
django-extensions==3.2.3
django-heroku==0.3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
flake8==7.1.1
gunicorn==23.0.0
inflection==0.5.1
isort==6.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.10
psycopg2-binary==2.9.3
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rpds-py==0.22.3
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.9.0
